requirements.txt
amqp==2.5.2
asgiref==3.2.3
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
billiard==3.6.1.0
blinker==1.4
celery==4.4.0
certifi==2018.1.18
cffi==1.13.2
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
Django==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==3.0.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==1.3.0
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
kombu==4.6.7
MarkupSafe==1.0
more-itertools==8.0.2
netifaces==0.10.4
oauthlib==2.0.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.19
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
pysqlite3==0.4.1
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
service-identity==16.0.0
six==1.11.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.22
vine==1.3.0
zipp==0.6.0
zope.interface==4.3.2
